+++ backend/src/relation/relation.service.ts
@@ -17,9 +17,6 @@ export class RelationService {
     const relations = await subjectUser.subjectRelations;
     const relation = relations.find(relation => relation.objectUser.userID === objectUser.userID);
 
-    if (!relation)
-      throw new NotFoundException('없는 User 관계입니다.');
-
     return (relation);
   }
 
@@ -27,11 +24,17 @@ export class RelationService {
     const { subjectUser, objectUser, relationType } = relationDto;
     const relation = await this.getRelationByUsers(subjectUser, objectUser);
 
+    if (!relation)
+      throw new NotFoundException('없는 User 관계입니다.');
+
     return (this.relationRepository.updateRelation(relation, relationType));
   }
 
   async deleteRelation(subjectUser: User, objectUser: User) {
     const relation = await this.getRelationByUsers(subjectUser, objectUser);
+
+    if (!relation)
+      throw new NotFoundException('없는 User 관계입니다.');
 
     this.relationRepository.deleteRelation(relation.relationID);
   }
